git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2667 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/entities/whatssid/MusicInfo.java
@@ -18,16 +18,16 @@ import libsidutils.fingerprinting.rest.beans.MusicInfoBean;
 @Table(name = "MusicInfo", indexes = { @Index(columnList = "idMusicInfo", name = "idMusicInfo_UNIQUE", unique = true) })
 public class MusicInfo {
 
-	private long idMusicInfo;
+	private int idMusicInfo;
 
 	@Id
 	@GeneratedValue (strategy = GenerationType.IDENTITY)
 	@Column(name = "idMusicInfo", nullable = false)
-	public long getIdMusicInfo() {
+	public int getIdMusicInfo() {
 		return idMusicInfo;
 	}
 
-	public void setIdMusicInfo(long idMusicInfo) {
+	public void setIdMusicInfo(int idMusicInfo) {
 		this.idMusicInfo = idMusicInfo;
 	}
